requirements/base.txt
# Core Django
Django==5.0.1
django-environ==0.11.2

# Database
mysqlclient==2.2.1

# Forms & Validation
django-crispy-forms==2.1
crispy-bootstrap4==2024.1

# Images
Pillow==10.2.0

# Payment Gateways
razorpay==1.4.1
stripe==7.11.0

# API (Optional)
djangorestframework==3.14.0
django-filter==23.5

# Security
django-cors-headers==4.3.1

# Utilities
python-decouple==3.8
